+++ packages/vue/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @lottiefiles/dotlottie-vue
 
+## 0.4.17
+
+### Patch Changes
+
+- Updated dependencies [26f4be4]
+  - @lottiefiles/dotlottie-web@0.36.0
+
 ## 0.4.16
 
 ### Patch Changes